+++ src/constants.js
@@ -121,6 +121,40 @@ export const ipcEvents = deepFreeze({
      * - `fetching`: New fetching state.
      */
     CLOUD_FETCHING_CHANGE: 'cloudFetchingChange',
+
+    /**
+     * System power is being suspended.
+     *
+     * Signature: `(event: string) => void`
+     */
+    POWER_SUSPEND: 'powerSuspend',
+
+    /**
+     * System power has been restored.
+     *
+     * Signature: `(event: string) => void`
+     */
+    POWER_RESUME: 'powerResume',
+
+    /**
+     * Network connection has dropped.
+     *
+     * NOTE: This is not 100% reliable because it's not always possible for Electron
+     *  to detect when the network is offline. For example, if the app is running in
+     *  a VM with its virtual network adapter configured to be "always on", yet the
+     *  host's network connection is dropped, the app will appear online when in fact
+     *  it's not.
+     *
+     * Signature: `(event: string) => void`
+     */
+    NETWORK_OFFLINE: 'networkOffline',
+
+    /**
+     * Network connection has been restored.
+     *
+     * Signature: `(event: string) => void`
+     */
+    NETWORK_ONLINE: 'networkOnline',
   },
 
   /** Invoked on the `main` process only. Returns a promise to be awaited. */